+++ include/GooseFEM/ElementCohesive8.h
@@ -230,10 +230,10 @@ public:
         xt::pytensor<double, 3> q_delta_u_global = xt::zeros<double>(this->shape_qvector());
 
     #pragma omp parallel for
-        for (size_t e = 0; e < m_nelem; ++e) {
+        for (ptrdiff_t e = 0; e < (ptrdiff_t)m_nelem; ++e) {
             auto u = xt::adapt(&elem_u(e, 0, 0), xt::xshape<s_nne, s_ndim>());
 
-            for (size_t q = 0; q < m_nip; ++q) {
+            for (ptrdiff_t q = 0; q < (ptrdiff_t)m_nip; ++q) {
                 // Interpolate displacements for lower surface nodes (0,1,2,3)
                 array_type::tensor<double, 1> u_lower = xt::empty<double>({s_ndim});
                 u_lower(0) = m_N(q, 0) * u(0, 0) + m_N(q, 1) * u(1, 0) + m_N(q, 2) * u(2, 0) + m_N(q, 3) * u(3, 0);
@@ -257,15 +257,15 @@ public:
         rot_mat.fill(0.0);
 
     #pragma omp parallel for
-        for (size_t e = 0; e < m_nelem; ++e) {
-            for (size_t q = 0; q < m_nip; ++q) {
+        for (ptrdiff_t e = 0; e < (ptrdiff_t)m_nelem; ++e) {
+            for (ptrdiff_t q = 0; q < (ptrdiff_t)m_nip; ++q) {
                 auto delta_u = xt::adapt(&q_delta_u_global(e, q, 0), xt::xshape<s_ndim>());
                 auto rot = xt::adapt(&m_rotation_matrix(e, q, 0, 0), xt::xshape<s_ndim, s_ndim>());
                 auto rot_output = xt::adapt(&rot_mat(e, q, 0, 0), xt::xshape<s_ndim, s_ndim>());
 
                 // Apply rotation: q_delta_u_norm_tan = R * delta_u_global
-                for (size_t i = 0; i < s_ndim; ++i) { 
-                    for (size_t j = 0; j < s_ndim; ++j) { 
+                for (ptrdiff_t i = 0; i < (ptrdiff_t)s_ndim; ++i) { 
+                    for (ptrdiff_t j = 0; j < (ptrdiff_t)s_ndim; ++j) { 
                         q_delta_u_norm_tan(e, q, i) += rot(i, j) * delta_u(j);
                     }
                 }
@@ -292,10 +292,10 @@ public:
         elem_f.fill(0.0);
 
     #pragma omp parallel for
-        for (size_t e = 0; e < m_nelem; ++e) {
+        for (ptrdiff_t e = 0; e < (ptrdiff_t)m_nelem; ++e) {
             auto f = xt::adapt(&elem_f(e, 0, 0), xt::xshape<s_nne, s_ndim>());
 
-            for (size_t q = 0; q < m_nip; ++q) {
+            for (ptrdiff_t q = 0; q < (ptrdiff_t)m_nip; ++q) {
                 auto tractions = xt::adapt(&q_tractions(e, q, 0), xt::xshape<s_ndim>());
                 auto& dA = m_vol(e, q);
 
@@ -305,7 +305,7 @@ public:
                 // This is implemented by summing contributions from each node.
                 
                 // For each of the 4 pairs of nodes (0-4, 1-5, 2-6, 3-7)
-                for (size_t pair_idx = 0; pair_idx < 4; ++pair_idx) {
+                for (ptrdiff_t pair_idx = 0; pair_idx < 4; ++pair_idx) {
                     size_t bottom_node_idx = pair_idx;
                     size_t top_node_idx = pair_idx + 4;
                     
@@ -315,7 +315,7 @@ public:
                     double N_top = m_N(q, top_node_idx);
                     
                     // For each spatial dimension (x, y, z)
-                    for (size_t dim = 0; dim < s_ndim; ++dim) {
+                    for (ptrdiff_t dim = 0; dim < (ptrdiff_t)s_ndim; ++dim) {
                         // The force on the bottom face is negative.
                         f(bottom_node_idx, dim) += -N_bottom * tractions(dim) * dA;
                         
@@ -370,10 +370,10 @@ public:
         elem_K.fill(0.0);
 
     #pragma omp parallel for
-        for (size_t e = 0; e < m_nelem; ++e) {
+        for (ptrdiff_t e = 0; e < (ptrdiff_t)m_nelem; ++e) {
             auto K_elem = xt::adapt(&elem_K(e, 0, 0), xt::xshape<s_nne * s_ndim, s_nne * s_ndim>());
 
-            for (size_t q = 0; q < m_nip; ++q) {
+            for (ptrdiff_t q = 0; q < (ptrdiff_t)m_nip; ++q) {
                 auto D_T_global = xt::adapt(&q_tangent_stiffness_global(e, q, 0, 0), xt::xshape<s_ndim, s_ndim>());
                 auto& dA = m_vol(e, q);
 
@@ -388,11 +388,11 @@ public:
                 // The separation is computed as the displacement of the top nodes minus the displacement of the corresponding bottom nodes.
                 // For example, top node 4 corresponds to bottom node 0.
 
-                for (size_t pair_idx = 0; pair_idx < 4; ++pair_idx) {
+                for (ptrdiff_t pair_idx = 0; pair_idx < 4; ++pair_idx) {
                     size_t bottom_node_idx = pair_idx;
                     size_t top_node_idx = pair_idx + 4;
 
-                    for (size_t dim = 0; dim < s_ndim; ++dim) {
+                    for (ptrdiff_t dim = 0; dim < (ptrdiff_t)s_ndim; ++dim) {
                         // Contribution from the bottom nodes (negative sign)
                         B_coh(dim, bottom_node_idx * s_ndim + dim) = -m_N(q, bottom_node_idx);
                         
@@ -408,18 +408,18 @@ public:
                 xt::xtensor_fixed<double, xt::xshape<s_ndim, s_nne * s_ndim>> Temp;
                 Temp.fill(0.0);
 
-                for (size_t i = 0; i < s_ndim; ++i) { 
-                    for (size_t k = 0; k < s_nne * s_ndim; ++k) { 
-                        for (size_t j = 0; j < s_ndim; ++j) { 
+                for (ptrdiff_t i = 0; i < (ptrdiff_t)s_ndim; ++i) { 
+                    for (ptrdiff_t k = 0; k < (ptrdiff_t)s_nne * s_ndim; ++k) { 
+                        for (ptrdiff_t j = 0; j < (ptrdiff_t)s_ndim; ++j) { 
                             Temp(i, k) += D_T_global(i, j) * B_coh(j, k);
                         }
                     }
                 }
 
-                for (size_t m = 0; m < s_nne * s_ndim; ++m) { 
-                    for (size_t n = 0; n < s_nne * s_ndim; ++n) { 
+                for (ptrdiff_t m = 0; m < (ptrdiff_t)s_nne * s_ndim; ++m) { 
+                    for (ptrdiff_t n = 0; n < (ptrdiff_t)s_nne * s_ndim; ++n) { 
                         double sum_val = 0.0;
-                        for (size_t p = 0; p < s_ndim; ++p) { 
+                        for (ptrdiff_t p = 0; p < (ptrdiff_t)s_ndim; ++p) { 
                             sum_val += B_coh(p, m) * Temp(p, n);
                         }
                         K_elem(m, n) += sum_val * dA;
@@ -459,15 +459,15 @@ private:
             array_type::tensor<double, 2> J_surface = xt::empty<double>({s_ndim, s_ndimxi}); 
 
     #pragma omp for
-            for (size_t e = 0; e < m_nelem; ++e) {
+            for (ptrdiff_t e = 0; e < (ptrdiff_t)m_nelem; ++e) {
                 auto x = xt::adapt(&m_x(e, 0, 0), xt::xshape<s_nne, s_ndim>());
 
-                for (size_t q = 0; q < m_nip; ++q) {
+                for (ptrdiff_t q = 0; q < (ptrdiff_t)m_nip; ++q) {
 
                     J_surface.fill(0.0);
 
                     // Loop over all 8 nodes
-                    for (size_t node_idx = 0; node_idx < s_nne; ++node_idx) {
+                    for (ptrdiff_t node_idx = 0; node_idx < (ptrdiff_t)s_nne; ++node_idx) {
                         // dX/dxi, dY/dxi, dZ/dxi 
                         J_surface(0, 0) += m_dNxi(q, node_idx, 0) * x(node_idx, 0);
                         J_surface(1, 0) += m_dNxi(q, node_idx, 0) * x(node_idx, 1);
